backend/requirements.txt
# ===============================
# Core Framework
# ===============================
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ===============================
# Configuration & Validation
# ===============================
pydantic==1.10.13
python-dotenv==1.0.1

# ===============================
# Authentication & Security
# ===============================
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# ===============================
# Database (PostgreSQL)
# ===============================
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# ===============================
# File Upload & Parsing
# ===============================
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.0

# ===============================
# OCR (Scanned PDFs)
# ===============================
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0

# ===============================
# AI / NLP (Governed Usage)
# ===============================
torch==2.2.1
transformers==4.37.2
sentence-transformers==2.3.1
langdetect==1.0.9

# ===============================
# Testing
# ===============================
pytest==8.0.0
httpx==0.26.0

# ===============================
# Utilities
# ===============================
python-dateutil==2.8.2
